calls/test/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
mcp==1.8.1
openai>=1.0.0
packaging==25.0
pluggy==1.5.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytest==7.4.3
python-dotenv==1.0.0
